AE-1987: Implement new ui for käskypäätös / varsinainen päätöstoimenpide
- All osapuoli specific fields are under the heading with the osapuoli name
- Separate osapuolitable removed as unnecessary
