+++ src/language/fi.json
@@ -1934,6 +1934,9 @@
         "decision-order-actual-decision": {
           "title": "Käskypäätös / varsinainen päätös",
           "info": "Rakennuksen omistaja ei ole vastannut kuulemiskirjeeseen tai perustellut riittävästi energiatodistuksen puuttumista, joten käsketään omistajaa hankkimaan energiatodistus ja tehdään käskypäätös uhkasakon asettamisesta. Käskypäätöksen allekirjoitus tapahtuu tämän järjestelmän ulkopuolella.",
+          "osapuoli-specific-data": "Osapuolikohtaiset tiedot",
+          "document": "Osapuolelle toimitetaan dokumentti",
+          "preview-button": "Esikatsele osapuolen dokumentti",
           "hearing-letter-answered": {
             "label": "Omistaja vastasi kuulemiskirjeeseen",
             "true": "Asianosainen antoi vastineen kuulemiskirjeeseen",
