+++ src/language/sv.json
@@ -1917,6 +1917,9 @@
         "decision-order-actual-decision": {
           "title": "Käskypäätös / varsinainen päätös (sv)",
           "info": "Rakennuksen omistaja ei ole vastannut kuulemiskirjeeseen tai perustellut riittävästi energiatodistuksen puuttumista, joten käsketään omistajaa hankkimaan energiatodistus ja tehdään käskypäätös uhkasakon asettamisesta. Käskypäätöksen allekirjoitus tapahtuu tämän järjestelmän ulkopuolella. (sv)",
+          "osapuoli-specific-data": "Osapuolikohtaiset tiedot (sv)",
+          "document": "Osapuolelle toimitetaan dokumentti (sv)",
+          "preview-button": "Esikatsele osapuolen dokumentti (sv)",
           "hearing-letter-answered": {
             "label": "Omistaja vastasi kuulemiskirjeeseen (sv)",
             "true": "Kyllä (sv)",
