+++ src/pages/valvonta-kaytto/new-actual-decision-toimenpide.svelte
@@ -7,20 +7,48 @@
   import Input from '@Component/Input/Input';
   import * as Osapuolet from '@Pages/valvonta-kaytto/osapuolet';
   import H2 from '@Component/H/H2';
+  import H3 from '@Component/H/H3';
   import * as Toimenpiteet from '@Pages/valvonta-kaytto/toimenpiteet';
+  import Checkbox from '@Component/Checkbox/Checkbox.svelte';
+  import * as Locales from '@Language/locale-utils';
+  import { locale } from '@Language/i18n';
+  import * as Selects from '@Component/Select/select-util';
+  import { isValid } from '@Utility/classification';
+  import TextButton from '@Component/Button/TextButton';
+  import * as ValvontaApi from '@Pages/valvonta-kaytto/valvonta-api';
 
+  export let id;
   export let toimenpide;
   export let henkilot;
   export let yritykset;
+  export let hallintoOikeudet = [];
+
+  export let preview;
+  export let previewPending;
+  export let disabled;
   export let text;
   export let i18n;
   export let schema;
 
-  const formatOsapuoliName = osapuoli =>
-    osapuoli.nimi || `${osapuoli.etunimi} ${osapuoli.sukunimi}`;
+  const types = {
+    yritys: {
+      label: yritys => yritys.nimi,
+      preview: ValvontaApi.previewToimenpideForYritysOsapuoli
+    },
+    henkilo: {
+      label: henkilo => `${henkilo.etunimi} ${henkilo.sukunimi}`,
+      preview: ValvontaApi.previewToimenpideForHenkiloOsapuoli
+    }
+  };
 
   let osapuolet = R.sort(R.ascend(R.prop('toimitustapa-id')))(
-    R.filter(Osapuolet.isOmistaja, R.concat(henkilot, yritykset))
+    R.filter(
+      Osapuolet.isOmistaja,
+      R.concat(
+        R.map(R.assoc('type', types.henkilo), henkilot),
+        R.map(R.assoc('type', types.yritys), yritykset)
+      )
+    )
   );
 
   const courtDataIndexForOsapuoli = osapuoliId =>
@@ -82,131 +110,197 @@
       {i18n} />
   </div>
 
-  {#each osapuolet as osapuoli}
-    <H2 text={formatOsapuoliName(osapuoli)} />
-
-    <div class="w-full py-4">
-      <Select2
-        bind:model={toimenpide}
-        lens={R.lensPath([
-          'type-specific-data',
-          'osapuoli-specific-data',
-          courtDataIndexForOsapuoli(osapuoli.id),
-          'recipient-answered'
-        ])}
-        format={value => text(toimenpide, `hearing-letter-answered.${value}`)}
-        label={text(toimenpide, 'hearing-letter-answered.label')}
-        required
-        items={[true, false]} />
-    </div>
-
-    {#if R.path(['type-specific-data', 'osapuoli-specific-data', courtDataIndexForOsapuoli(osapuoli.id), 'recipient-answered'], toimenpide)}
-      <div class="w-full py-4">
-        <Textarea
-          id={'toimenpide.answer-commentary-fi'}
-          name={'toimenpide.answer-commentary-fi'}
-          label={text(toimenpide, 'answer-commentary-fi')}
+  <div class="w-full py-4">
+    <H2 text={text(toimenpide, 'osapuoli-specific-data')} />
+    {#each osapuolet as osapuoli}
+      <div class="py-4">
+        <H3 text={osapuoli.type.label(osapuoli)} />
+        <Checkbox
           bind:model={toimenpide}
+          label={text(toimenpide, 'document')}
           lens={R.lensPath([
             'type-specific-data',
             'osapuoli-specific-data',
             courtDataIndexForOsapuoli(osapuoli.id),
-            'answer-commentary-fi'
-          ])}
-          required
-          format={Maybe.orSome('')}
-          parse={Parsers.optionalString}
-          validators={R.path(
-            [
-              'type-specific-data',
-              'osapuoli-specific-data',
-              courtDataIndexForOsapuoli(osapuoli.id),
-              'answer-commentary-fi'
-            ],
-            schema
-          )}
-          {i18n} />
-      </div>
-      <div class="w-full py-4">
-        <Textarea
-          id={'toimenpide.answer-commentary-sv'}
-          name={'toimenpide.answer-commentary-sv'}
-          label={text(toimenpide, 'answer-commentary-sv')}
-          bind:model={toimenpide}
-          lens={R.lensPath([
-            'type-specific-data',
-            'osapuoli-specific-data',
-            courtDataIndexForOsapuoli(osapuoli.id),
-            'answer-commentary-sv'
-          ])}
-          required
-          format={Maybe.orSome('')}
-          parse={Parsers.optionalString}
-          validators={R.path(
-            [
-              'type-specific-data',
-              'osapuoli-specific-data',
-              courtDataIndexForOsapuoli(osapuoli.id),
-              'answer-commentary-sv'
-            ],
-            schema
-          )}
-          {i18n} />
-      </div>
+            'document'
+          ])} />
+        {#if Toimenpiteet.documentExistsForOsapuoli(toimenpide, osapuoli.id)}
+          <div class="w-full py-4">
+            <Select2
+              bind:model={toimenpide}
+              lens={R.lensPath([
+                'type-specific-data',
+                'osapuoli-specific-data',
+                courtDataIndexForOsapuoli(osapuoli.id),
+                'recipient-answered'
+              ])}
+              format={value =>
+                text(toimenpide, `hearing-letter-answered.${value}`)}
+              label={text(toimenpide, 'hearing-letter-answered.label')}
+              required
+              items={[true, false]} />
+          </div>
 
-      <div class="w-full py-4">
-        <Textarea
-          id={'toimenpide.statement-fi'}
-          name={'toimenpide.statement-fi'}
-          label={text(toimenpide, 'statement-fi')}
-          bind:model={toimenpide}
-          lens={R.lensPath([
-            'type-specific-data',
-            'osapuoli-specific-data',
-            courtDataIndexForOsapuoli(osapuoli.id),
-            'statement-fi'
-          ])}
-          required
-          format={Maybe.orSome('')}
-          parse={Parsers.optionalString}
-          validators={R.path(
-            [
-              'type-specific-data',
-              'osapuoli-specific-data',
-              courtDataIndexForOsapuoli(osapuoli.id),
-              'statement-fi'
-            ],
-            schema
-          )}
-          {i18n} />
-      </div>
+          {#if R.path(['type-specific-data', 'osapuoli-specific-data', courtDataIndexForOsapuoli(osapuoli.id), 'recipient-answered'], toimenpide)}
+            <div class="w-full py-4">
+              <Textarea
+                id={'toimenpide.answer-commentary-fi'}
+                name={'toimenpide.answer-commentary-fi'}
+                label={text(toimenpide, 'answer-commentary-fi')}
+                bind:model={toimenpide}
+                lens={R.lensPath([
+                  'type-specific-data',
+                  'osapuoli-specific-data',
+                  courtDataIndexForOsapuoli(osapuoli.id),
+                  'answer-commentary-fi'
+                ])}
+                required
+                format={Maybe.orSome('')}
+                parse={Parsers.optionalString}
+                validators={R.path(
+                  [
+                    'type-specific-data',
+                    'osapuoli-specific-data',
+                    courtDataIndexForOsapuoli(osapuoli.id),
+                    'answer-commentary-fi'
+                  ],
+                  schema
+                )}
+                {i18n} />
+            </div>
+            <div class="w-full py-4">
+              <Textarea
+                id={'toimenpide.answer-commentary-sv'}
+                name={'toimenpide.answer-commentary-sv'}
+                label={text(toimenpide, 'answer-commentary-sv')}
+                bind:model={toimenpide}
+                lens={R.lensPath([
+                  'type-specific-data',
+                  'osapuoli-specific-data',
+                  courtDataIndexForOsapuoli(osapuoli.id),
+                  'answer-commentary-sv'
+                ])}
+                required
+                format={Maybe.orSome('')}
+                parse={Parsers.optionalString}
+                validators={R.path(
+                  [
+                    'type-specific-data',
+                    'osapuoli-specific-data',
+                    courtDataIndexForOsapuoli(osapuoli.id),
+                    'answer-commentary-sv'
+                  ],
+                  schema
+                )}
+                {i18n} />
+            </div>
 
-      <div class="w-full py-4">
-        <Textarea
-          id={'toimenpide.statement-sv'}
-          name={'toimenpide.statement-sv'}
-          label={text(toimenpide, 'statement-sv')}
-          bind:model={toimenpide}
-          lens={R.lensPath([
-            'type-specific-data',
-            'osapuoli-specific-data',
-            courtDataIndexForOsapuoli(osapuoli.id),
-            'statement-sv'
-          ])}
-          required
-          format={Maybe.orSome('')}
-          parse={Parsers.optionalString}
-          validators={R.path(
-            [
-              'type-specific-data',
-              'osapuoli-specific-data',
-              courtDataIndexForOsapuoli(osapuoli.id),
-              'statement-sv'
-            ],
-            schema
-          )}
-          {i18n} />
+            <div class="w-full py-4">
+              <Textarea
+                id={'toimenpide.statement-fi'}
+                name={'toimenpide.statement-fi'}
+                label={text(toimenpide, 'statement-fi')}
+                bind:model={toimenpide}
+                lens={R.lensPath([
+                  'type-specific-data',
+                  'osapuoli-specific-data',
+                  courtDataIndexForOsapuoli(osapuoli.id),
+                  'statement-fi'
+                ])}
+                required
+                format={Maybe.orSome('')}
+                parse={Parsers.optionalString}
+                validators={R.path(
+                  [
+                    'type-specific-data',
+                    'osapuoli-specific-data',
+                    courtDataIndexForOsapuoli(osapuoli.id),
+                    'statement-fi'
+                  ],
+                  schema
+                )}
+                {i18n} />
+            </div>
+
+            <div class="w-full py-4">
+              <Textarea
+                id={'toimenpide.statement-sv'}
+                name={'toimenpide.statement-sv'}
+                label={text(toimenpide, 'statement-sv')}
+                bind:model={toimenpide}
+                lens={R.lensPath([
+                  'type-specific-data',
+                  'osapuoli-specific-data',
+                  courtDataIndexForOsapuoli(osapuoli.id),
+                  'statement-sv'
+                ])}
+                required
+                format={Maybe.orSome('')}
+                parse={Parsers.optionalString}
+                validators={R.path(
+                  [
+                    'type-specific-data',
+                    'osapuoli-specific-data',
+                    courtDataIndexForOsapuoli(osapuoli.id),
+                    'statement-sv'
+                  ],
+                  schema
+                )}
+                {i18n} />
+            </div>
+          {/if}
+          <div class="w-full py-4">
+            <Select2
+              bind:model={toimenpide}
+              label="Hallinto-oikeus"
+              required
+              lens={R.lensPath([
+                'type-specific-data',
+                'osapuoli-specific-data',
+                courtDataIndexForOsapuoli(osapuoli.id),
+                'hallinto-oikeus-id'
+              ])}
+              modelToItem={Maybe.fold(
+                Maybe.None(),
+                Maybe.findById(R.__, hallintoOikeudet)
+              )}
+              itemToModel={Maybe.fold(Maybe.None(), it => Maybe.Some(it.id))}
+              format={Maybe.fold(
+                i18n('validation.no-selection'),
+                Locales.label($locale)
+              )}
+              validators={R.path(
+                [
+                  'type-specific-data',
+                  'osapuoli-specific-data',
+                  courtDataIndexForOsapuoli(osapuoli.id),
+                  'hallinto-oikeus-id'
+                ],
+                schema
+              )}
+              items={Selects.addNoSelection(
+                R.filter(isValid, hallintoOikeudet)
+              )} />
+          </div>
+          <div class="w-full py-6">
+            <TextButton
+              text={text(toimenpide, 'preview-button')}
+              icon="visibility"
+              type="button"
+              on:click={_ =>
+                preview(osapuoli.type.preview(id, osapuoli.id, toimenpide))}
+              {disabled} />
+          </div>
+        {/if}
       </div>
-    {/if}
-  {/each}
+    {/each}
+  </div>
+
+  <div class="w-full py-4">
+    <p>
+      <span class="font-icon mr-1">info</span>{i18n(
+        'valvonta.kaytto.toimenpide.manually-sent'
+      )}
+    </p>
+  </div>
 </div>
